requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
boto3==1.28.46
botocore==1.31.46
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.5.14
dnspython==2.4.2
eventlet==0.33.3
exceptiongroup==1.1.3
fastapi==0.99.1
filelock==3.12.3
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
langchain==0.0.285
langsmith==0.0.35
llama-index==0.8.22
loguru==0.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.28.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pinecone-client==2.2.2
posthog==3.0.2
protobuf==4.21.12
psycopg2-binary==2.9.7
pulsar-client==3.3.0
pydantic==1.10.12
pydantic_core==2.6.3
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.0
python-socketio==5.9.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
s3transfer==0.6.2
safetensors==0.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.66.1
transformers==4.33.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
